+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Update `MLBProbableStartingPitcher` fields (#77)
+- Update `MLBProbableStartingPitcher` composite key from `EventID` to `EventID` and `TeamID` (#77)
+### Fixed
+- Skip TBD probable starting pitchers (#77)
 
 ## [0.9.0] - 2025-05-27
 ### Added